Citation for a survey

Gets a full citation for a survey().

Usage

get_citation(x)

Arguments

x

a character vector of surveys to cite

Value

citation as bibentry

Examples

data(polymod)
citation <- get_citation(polymod)
#> Using POLYMOD social contact data. To cite this in a publication, use the 'get_citation()' function
print(citation)
#> Mossong J, Hens N, Jit M, Beutels P, Auranen K, Mikolajczyk R, Massari
#> M, Salmaso S, Tomba GS, Wallinga J, Heijne J, Sadkowska-Todys M,
#> Rosinska M, Edmunds WJ (2017). “POLYMOD social contact data.”
#> doi:10.5281/zenodo.1157934 <https://doi.org/10.5281/zenodo.1157934>,
#> Version 1.1.
print(citation, style = "bibtex")
#> @Misc{,
#>   title = {POLYMOD social contact data},
#>   author = {Joel Mossong and Niel Hens and Mark Jit and Philippe Beutels and Kari Auranen and Rafael Mikolajczyk and Marco Massari and Stefania Salmaso and Gianpaolo Scalia Tomba and Jacco Wallinga and Janneke Heijne and Malgorzata Sadkowska-Todys and Magdalena Rosinska and W. John Edmunds},
#>   doi = {https://doi.org/10.5281/zenodo.1157934},
#>   note = {Version 1.1},
#>   year = {2017},
#> }