List all countries and regions for which socialmixr has population data

Uses the World Population Prospects data from the wpp2017 package

Usage

wpp_countries()

Value

list of countries

Examples

wpp_countries()
#>   [1] "China"                            "Burundi"                         
#>   [3] "Comoros"                          "Djibouti"                        
#>   [5] "Eritrea"                          "Ethiopia"                        
#>   [7] "Kenya"                            "Madagascar"                      
#>   [9] "Malawi"                           "Mauritius"                       
#>  [11] "Mayotte"                          "Mozambique"                      
#>  [13] "Réunion"                          "Rwanda"                          
#>  [15] "Seychelles"                       "Somalia"                         
#>  [17] "South Sudan"                      "Uganda"                          
#>  [19] "Tanzania"                         "Zambia"                          
#>  [21] "Zimbabwe"                         "Angola"                          
#>  [23] "Cameroon"                         "Central African Republic"        
#>  [25] "Chad"                             "Congo - Brazzaville"             
#>  [27] "Congo - Kinshasa"                 "Equatorial Guinea"               
#>  [29] "Gabon"                            "São Tomé & Príncipe"             
#>  [31] "Algeria"                          "Egypt"                           
#>  [33] "Libya"                            "Morocco"                         
#>  [35] "Sudan"                            "Tunisia"                         
#>  [37] "Western Sahara"                   "Botswana"                        
#>  [39] "Lesotho"                          "Namibia"                         
#>  [41] "South Africa"                     "Eswatini"                        
#>  [43] "Benin"                            "Burkina Faso"                    
#>  [45] "Cape Verde"                       "Côte d’Ivoire"                   
#>  [47] "Gambia"                           "Ghana"                           
#>  [49] "Guinea"                           "Guinea-Bissau"                   
#>  [51] "Liberia"                          "Mali"                            
#>  [53] "Mauritania"                       "Niger"                           
#>  [55] "Nigeria"                          "Senegal"                         
#>  [57] "Sierra Leone"                     "Togo"                            
#>  [59] "China"                            "Hong Kong SAR China"             
#>  [61] "Macao SAR China"                  "Taiwan"                          
#>  [63] "North Korea"                      "Japan"                           
#>  [65] "Mongolia"                         "South Korea"                     
#>  [67] "Kazakhstan"                       "Kyrgyzstan"                      
#>  [69] "Tajikistan"                       "Turkmenistan"                    
#>  [71] "Uzbekistan"                       "Afghanistan"                     
#>  [73] "Bangladesh"                       "Bhutan"                          
#>  [75] "India"                            "Iran"                            
#>  [77] "Maldives"                         "Nepal"                           
#>  [79] "Pakistan"                         "Sri Lanka"                       
#>  [81] "Brunei"                           "Cambodia"                        
#>  [83] "Indonesia"                        "Laos"                            
#>  [85] "Malaysia"                         "Myanmar (Burma)"                 
#>  [87] "Philippines"                      "Singapore"                       
#>  [89] "Thailand"                         "Timor-Leste"                     
#>  [91] "Vietnam"                          "Armenia"                         
#>  [93] "Azerbaijan"                       "Bahrain"                         
#>  [95] "Cyprus"                           "Georgia"                         
#>  [97] "Iraq"                             "Israel"                          
#>  [99] "Jordan"                           "Kuwait"                          
#> [101] "Lebanon"                          "Oman"                            
#> [103] "Qatar"                            "Saudi Arabia"                    
#> [105] "Palestinian Territories"          "Syria"                           
#> [107] "Turkey"                           "United Arab Emirates"            
#> [109] "Yemen"                            "Belarus"                         
#> [111] "Bulgaria"                         "Czechia"                         
#> [113] "Hungary"                          "Poland"                          
#> [115] "Moldova"                          "Romania"                         
#> [117] "Russia"                           "Slovakia"                        
#> [119] "Ukraine"                          "Channel Islands"                 
#> [121] "Denmark"                          "Estonia"                         
#> [123] "Finland"                          "Iceland"                         
#> [125] "Ireland"                          "Latvia"                          
#> [127] "Lithuania"                        "Norway"                          
#> [129] "Sweden"                           "United Kingdom"                  
#> [131] "Albania"                          "Bosnia & Herzegovina"            
#> [133] "Croatia"                          "Greece"                          
#> [135] "Italy"                            "Malta"                           
#> [137] "Montenegro"                       "Portugal"                        
#> [139] "Serbia"                           "Slovenia"                        
#> [141] "Spain"                            "North Macedonia"                 
#> [143] "Austria"                          "Belgium"                         
#> [145] "France"                           "Germany"                         
#> [147] "Luxembourg"                       "Netherlands"                     
#> [149] "Switzerland"                      "Antigua & Barbuda"               
#> [151] "Aruba"                            "Bahamas"                         
#> [153] "Barbados"                         "Cuba"                            
#> [155] "Curaçao"                          "Dominican Republic"              
#> [157] "Grenada"                          "Guadeloupe"                      
#> [159] "Haiti"                            "Jamaica"                         
#> [161] "Martinique"                       "Puerto Rico"                     
#> [163] "St. Lucia"                        "St. Vincent & Grenadines"        
#> [165] "Trinidad & Tobago"                "U.S. Virgin Islands"             
#> [167] "Belize"                           "Costa Rica"                      
#> [169] "El Salvador"                      "Guatemala"                       
#> [171] "Honduras"                         "Mexico"                          
#> [173] "Nicaragua"                        "Panama"                          
#> [175] "Argentina"                        "Bolivia"                         
#> [177] "Brazil"                           "Chile"                           
#> [179] "Colombia"                         "Ecuador"                         
#> [181] "French Guiana"                    "Guyana"                          
#> [183] "Paraguay"                         "Peru"                            
#> [185] "Suriname"                         "Uruguay"                         
#> [187] "Venezuela"                        "Canada"                          
#> [189] "United States"                    "Australia"                       
#> [191] "New Zealand"                      "Fiji"                            
#> [193] "New Caledonia"                    "Papua New Guinea"                
#> [195] "Solomon Islands"                  "Vanuatu"                         
#> [197] "Guam"                             "Kiribati"                        
#> [199] "Micronesia (Federated States of)" "French Polynesia"                
#> [201] "Samoa"                            "Tonga"